requirements.txt
python-telegram-bot>=20.0
beautifulsoup4>=4.12.0
httpx>=0.24.0
cachetools>=5.3.0
python-dateutil>=2.8.2
Pillow>=10.0.0
lxml>=4.9.0
reportlab>=4.0.0

